+++ plugins/@grouparoo/app-templates/src/source/query/profiles.ts
